[infra] Enable dart.use_update_scripts experiment on four builders
The new result update scripts, that write to the staging Firestore
database, are behind a buildbucket experiment flag in the recipes.

Enable them on the builders (and try builders)
      analyzer-win-release
      pkg-win-release
      pkg-mac-release-arm64
      vm-kernel-checked-linux-release-x64

Later, they will be enabled on all builders, and switched to write
to the production Firestore database.

Bug: b/205520378
Change-Id: I46bbd30cba3bb5001385f87748a8f4a81477317e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/220181
Commit-Queue: William Hesse <whesse@google.com>
Reviewed-by: Alexander Thomas <athom@google.com>