[cfe] Compute macro declarations/applications during buildOutline
This CL moves the computation of macro declarations/applications
from macro_test to KernelTarget.buildOutline. Initially these are
computed as late in the phase as is convenient. Later these
computations are intended to be moved as early as possible to ensure
a performant implementation of macros.

Change-Id: Ia2255937eba8d87b6ca8f3f03d13960206f176d3
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/220765
Commit-Queue: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Jens Johansen <jensj@google.com>