Another work around for an invalidation issue.
I still don't know why it happens.
But there might be something else as well, as users complaining
about the server not responsive. And one suspect is this infinite
cycle that I added. I thought it is safe, but something is wrong,
and it could be it.

Change-Id: I6095e1493c4cc49e6a5540bf30d244747c36641c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/221023
Reviewed-by: Jaime Wren <jwren@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>