Mention Object.hashCode in the note about LinkedHashSet in the Set documentation
The note about `LinkedHashSet` in the `Set` documentation is a little misleading, and it can lead developers to implement `==` without `hashCode`, only to be confused when their set doesn't work correctly.

See this [StackOverflow question](https://stackoverflow.com/questions/69664073/dart-comparison-of-objects-within-a-list/69694646#69694646), for example.

This PR mentions `Object.hashCode` along with `Object.==` in the `Set` documentation explaining `LinkedHashSet`.

Closes https://github.com/dart-lang/sdk/pull/47537
https://github.com/dart-lang/sdk/pull/47537

GitOrigin-RevId: 110a2d8181d4cdce90a576a7d0b5de523f0da6e6
Change-Id: Ic141383d5fa9fa2f9f733611df714041f783e1d2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/217861
Reviewed-by: Lasse R.H. Nielsen <lrn@google.com>
Commit-Queue: Michael Thomsen <mit@google.com>