[vm, gc] Fix TSAN failure caused by loading a TypedDataView's backing store's header.
This header is only used by an assert, but the load was not removed by the C++ compiler in release mode. Explicitly exclude the loads with the preprocessor so they only happen in debug mode, and limit the assert reading the backing store's header to non-parallel scavenges only.

TEST=tsan
Bug: https://github.com/dart-lang/sdk/issues/47895
Change-Id: Idbbdb743404d8ee15b0f8e13332c022453de085c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/223660
Reviewed-by: Alexander Aprelev <aam@google.com>
Commit-Queue: Ryan Macnak <rmacnak@google.com>