Use the full ResolverVisitor when testing ElementResolver and StaticTypeAnalyzer
I'm experimenting with refactoring some code in a way that changes the
interface between the ResolverVisitor and the two visitors that
invokes (ElementResolver and StaticTypeAnalyzer).  The division of
responsibility among these three classes is not very well-defined or
consistent due to their highly coupled nature and the fact that they
are always invoked in tandem.  This makes it difficult to refactor
them without causing churn on the ElementResolver and StaticAnalyzer
unit tests.

This change reduces the churn by changing the ElementResolver and
StaticAnalyzer unit tests to invoke the full ResolverVisitor; this
makes the interface among the three classes a fully private
implementation detail, so that we can adjust it without causing any
test breakages.

Change-Id: Ice4c0d98138e59bac52ed67eecf3ee535ca75e1b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/223141
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>