[vm/compiler] Handle yield continuations inside unreachable code
Previously, when concatenating two fragments of flow graph

  F1(entry1, end1)
  F2(entry2, end2)

we discarded F2 if F1 is closed. This is not correct
when F2 contains yield continuation point in the middle,
because later we add a branch targeting that continuation point,
making end2 reachable.

This change fixes concatenation of fragments to continue building
flow graph with the fragment

  F(entry1, end2)

even if F1 is closed (but we don't link end1 to entry2
in such case).

TEST=runtime/tests/vm/dart/regress_flutter66765_test.dart

Issue: https://github.com/flutter/flutter/issues/66765
Change-Id: I145cb4a8952e29665353688dccd232b3e9268878
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/224220
Reviewed-by: Ryan Macnak <rmacnak@google.com>
Reviewed-by: Martin Kustermann <kustermann@google.com>
Commit-Queue: Alexander Markov <alexmarkov@google.com>