[vm, lib] Refine List copy loops
`growableList.toList()` was 10-50% faster than `List.of(growableList)`.
This change erases the performance difference.

The loops were slightly different. The faster version created
the _List backing store for the list rather than assigning to the
_GrowableList.

It is not clear if the speedup is from an easier type test (the _List is
'raw'), or from a less indirect assignment, or from a fixed length list
having an an easier-to-optimize bounds check, or whether a non-zero loop
trip count enables any hoisting.

This change uses the faster pattern in the _GrowableList._ofXXXList
methods.

I tried to simplify `toList` to delegate to one of the methods
with the faster loop, but it was 2-5% slower on ListCopy.toList.2,
presumably the extra call is significant for a short list. If there
was an annotation for aggressive inlining into `toList` perhaps that
would close the gap.


dart-aot-null armv8c, 0.5 noise and over

ListCopy.List.num.from.2            24.17% (1.6 noise)
ListCopy.List.of.2                  28.08% (1.7 noise)
ListCopy.spread.num.2               28.73% (2.1 noise)
ListCopy.List.num.from.100          54.49% (1.9 noise)
ListCopy.List.of.100                55.55% (1.9 noise)
ListCopy.spread.num.100             55.77% (2.0 noise)


Change-Id: I7b780835faf50b673fd4b1a08a8c4e2d350d7d5b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/224780
Reviewed-by: Alexander Markov <alexmarkov@google.com>
Commit-Queue: Stephen Adams <sra@google.com>