[cfe]: Accept multiple libraries in compile_platform.
Allow a comma-separated list of dart:* libraries. This will be used in
other environments to include dart:ui when invoking compile_platform.

This is related to https://github.com/dart-lang/sdk/issues/30957.
Technically it doesn't remove our reliance on extraRequiredLibraries,
but it is the last blocking issue if we want to move in that direction.

This changes:
- command_line: to accept multiple inputs in compile_platform,
  adding more arguments was more complex and less backwards
  compatible than this change
- proessed_options: to not issue an error when all inputs
  provided are dart:* libraries if no `.packages` file was specified
  (which is the case for compile_plarform after this change).

Change-Id: Ibfaaa8ff216f6179cc4de7e2c66327d6ccde9b4d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/225100
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Sigmund Cherem <sigmund@google.com>