[cfe] Add MacroAnnotationParser to support early recognition
This CL adds a parser listener that supports (the initial) part of
the syntax needed to support detection of macro application early in
the pipeline. The listener is deliberately implementing (as opposed to
extending) the listener interface to ensure that all case handled in
the listener itself. The listener keeps track of whether unsupported
constructs have been seen to ensure that only valid syntax results
in a macro application.

Change-Id: I6742a0daa39c80f09759b41b9abdd7e693da4cc0
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/225723
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>