Partial fix for parsing explicitly typed super formal parameters.
I need at least this much to cover a few test cases in the analyzer,
but it might be not the best implementation. It seems to not break
anything, so maybe we can land this for now, while we are waiting for
the better fix of the referenced issue?

Bug: https://github.com/dart-lang/sdk/issues/47951
Change-Id: I428e32f4435b3229cb2dc41678182cde0e2f98e8
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/226200
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>