[vm/test] Fix vm/cc/IRTest_TypedDataAOT_FunctionalGetSet on arm
Make sure this unit test does not rely on AOT table dispatch as it is
not available in unit tests. This is done by avoiding 'operator ~/`
which is not optimized on certain ARMs and generated as a call which
needs dispatch table.

TEST=vm/cc/IRTest_TypedDataAOT_FunctionalGetSet
Fixes https://github.com/dart-lang/sdk/issues/48054

Change-Id: Icf102fbd01d7a479b5a79e6b840045642b20cc74
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/226300
Reviewed-by: Slava Egorov <vegorov@google.com>
Commit-Queue: Alexander Markov <alexmarkov@google.com>