[vm] Replace synthetic expression evaluation classes with real classes in types
When loading and compiling a function for expression evaluation
we might reference synthetic class of that function in types
(e.g. receiver type). However, that class is not registered
(so it could be garbage collected) and it has kIllegalCid.
Types constructed using such class are not very useful and
may crash if class is queried (for example if type is finalized).

This change attempts to replace such synthetic classes with
real classes in types constructed during loading.

TEST=ci

Issue: https://github.com/dart-lang/sdk/issues/48024
Change-Id: I6c5c1b1d20ac2ce4c43e5a7e456d5a8756a43727
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/226603
Reviewed-by: Ryan Macnak <rmacnak@google.com>
Commit-Queue: Alexander Markov <alexmarkov@google.com>