Flow analysis: add assertions to verify that API is used correctly
This change adds assertions to ensure that the first phase of flow
analysis (calls to the AssignedVariables class) is complete before the
FlowAnalysis class is created.  It also calls AssignedVariables.finish
(which does additional assertions) from the FlowAnalysis constructor
(if the client hasn't done so already).

Adding these assersions exposed a few minor (but benign) bugs in the
usage of the flow analysis API.

- The flow analysis unit tests were failing to call
  AssignedVariables.declare.  This should have been caught by
  AssignedVariables.finish, but the flow analysis unit tests were
  failing to call that as well.

- Several flow analysis unit tests failed to declare the variables
  they were using (or in one case, the declaration was at the wrong
  location).

- In the CFE, DeclaredSourceConstructorBuilder.buildOutlineExpressions
  was failing to set up scopes properly when compiling a const
  constructor.

- In the CFE, SourceLoader.buildExpression was failing to call
  AssignedVariables.declare for variables in scope.

- In the CFE, the TypeInferrerImpl constructor was constructing the
  FlowAnalysis object eagerly, even though the AssignedVariables phase
  often hadn't completed yet.  I've fixed this by constructing
  FlowAnalysis lazily.

A follow-up CL will take advantage of the invariants checked by these
assertions to introduce the logic necessary to fix issue #47991.

Change-Id: Id680020dcd525632afa0820bc49c7b966b1ea534
Bug: https://github.com/dart-lang/sdk/issues/47991
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/227320
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>