[cfe] Don't check bounds on uninferred invocations
Invocations that haven't been inferred are now not checked wrt. bounds.
These can occur in erroneous code or in outline expressions that are
currently created both in the outline phase (where they are fully
inferred) and in the body building phase where they are not inferred.

Additionally, an assertion is added to verify that these invocations
are no longer part of the AST, and therefore safe to leave incomplete.

Closes https://github.com/dart-lang/sdk/issues/48148

Change-Id: I6d83c2db05d43b016e21abf915cbe5a312b20c79
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/228560
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>