[vm/compiler] Ensure that right operand range is set for ShiftUint32Op created during range analysis
Code generation for shift instructions relies on the range information
for the right operand in order to generate shifts by constants and omit
certain runtime checks.

When range analysis creates ShiftUint32Op it has not been setting
range of the right operand, which resulted in suboptimal code.
This change fixes that by using BinaryIntegerOpInstr::Make which
initializes range of the right operand if it is a constant.

TEST=vm/cc/RangeAnalysis_ShiftUint32Op
Fixes https://github.com/dart-lang/sdk/issues/48153

Change-Id: I7df4bb86745bad8712862e3ea07c75b2c4898288
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/228684
Reviewed-by: Slava Egorov <vegorov@google.com>
Commit-Queue: Alexander Markov <alexmarkov@google.com>