[cfe] Compute computation layers from precompiled libraries
The previous computation of compilation layers used source/dill to
determine which macro libraries need to be precompiled. Since the
target of the macro runtime (the VM) and the output might not be the
same, we need to use the provided precompiled libraries to determine
which macros that need to be precompiled.

Change-Id: Id39e0cff5f85b4f37a16ac531d97c53f8971ecce
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/229140
Reviewed-by: Jens Johansen <jensj@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>